+++ examples/generators/production_python_smart_contract_python/pyproject.toml
@@ -10,7 +10,7 @@ python = "^3.12"
 algokit-utils = "^2.3.0"
 python-dotenv = "^1.0.0"
 algorand-python = "^1.0.0"
-algorand-python-testing = {git = "https://github.com/algorandfoundation/puya.git", rev = "feat/unit-testing", subdirectory = "algopy_testing"}
+algorand-python-testing = {git = "https://github.com/algorandfoundation/puya.git", rev = "fix/algopy_testing_refinements", subdirectory = "algopy_testing"}
 
 [tool.poetry.group.dev.dependencies]
 algokit-client-generator = "^1.1.3"
