+++ examples/generators/production_python_smart_contract_python/smart_contracts/__main__.py
@@ -4,16 +4,16 @@ from pathlib import Path
 
 from dotenv import load_dotenv
 
-from smart_contracts._helpers.build import build
-from smart_contracts._helpers.deploy import deploy
-from smart_contracts._helpers.util import find_app_spec_file
+from smart_contracts.config import contracts
+from smart_contracts.helpers.build import build
+from smart_contracts.helpers.deploy import deploy
+from smart_contracts.helpers.util import find_app_spec_file
 
 # Uncomment the following lines to enable auto generation of AVM Debugger compliant sourcemap and simulation trace file.
 # Learn more about using AlgoKit AVM Debugger to debug your TEAL source codes and inspect various kinds of
 # Algorand transactions in atomic groups -> https://github.com/algorandfoundation/algokit-avm-vscode-debugger
 # from algokit_utils.config import config
 # config.configure(debug=True, trace_all=True)
-from smart_contracts.config import contracts
 logging.basicConfig(
     level=logging.DEBUG, format="%(asctime)s %(levelname)-10s: %(message)s"
 )
