+++ examples/generators/production_python_smart_contract_python/smart_contracts/helpers/util.py
