+++ examples/generators/production_python_smart_contract_typescript/smart_contracts/helpers/util.py
