+++ examples/production_python/smart_contracts/helpers/util.py
