+++ examples/production_python/tests/hello_world_client_test.py
@@ -1,25 +1,48 @@
-from collections.abc import Generator
-
+import algokit_utils
 import pytest
-from algopy_testing import AlgopyTestContext, algopy_testing_context
+from algokit_utils import get_localnet_default_account
+from algokit_utils.config import config
+from algosdk.v2client.algod import AlgodClient
+from algosdk.v2client.indexer import IndexerClient
 
-from smart_contracts.hello_world.contract import HelloWorld
-
-
-@pytest.fixture()
-def context() -> Generator[AlgopyTestContext, None, None]:
-    with algopy_testing_context() as ctx:
-        yield ctx
-        ctx.reset()
+from smart_contracts.artifacts.hello_world.client import HelloWorldClient
 
 
-def test_hello(context: AlgopyTestContext) -> None:
-    # Arrange
-    dummy_input = context.any_string()
-    contract = HelloWorld()
+@pytest.fixture(scope="session")
+def hello_world_client(
+    algod_client: AlgodClient, indexer_client: IndexerClient
+) -> HelloWorldClient:
+    config.configure(
+        debug=True,
+        # trace_all=True,
+    )
 
-    # Act
-    output = contract.hello(dummy_input)
+    client = HelloWorldClient(
+        algod_client,
+        creator=get_localnet_default_account(algod_client),
+        indexer_client=indexer_client,
+    )
 
-    # Assert
-    assert output == f"Hello, {dummy_input}"
+    client.deploy(
+        on_schema_break=algokit_utils.OnSchemaBreak.AppendApp,
+        on_update=algokit_utils.OnUpdate.AppendApp,
+    )
+    return client
+
+
+def test_says_hello(hello_world_client: HelloWorldClient) -> None:
+    result = hello_world_client.hello(name="World")
+
+    assert result.return_value == "Hello, World"
+
+
+def test_simulate_says_hello_with_correct_budget_consumed(
+    hello_world_client: HelloWorldClient, algod_client: AlgodClient
+) -> None:
+    result = (
+        hello_world_client.compose().hello(name="World").hello(name="Jane").simulate()
+    )
+
+    assert result.abi_results[0].return_value == "Hello, World"
+    assert result.abi_results[1].return_value == "Hello, Jane"
+    assert result.simulate_response["txn-groups"][0]["app-budget-consumed"] < 100
