+++ examples/production_python/tests/hello_world_test.py
@@ -1,48 +1,25 @@
-import algokit_utils
+from collections.abc import Generator
+
 import pytest
-from algokit_utils import get_localnet_default_account
-from algokit_utils.config import config
-from algosdk.v2client.algod import AlgodClient
-from algosdk.v2client.indexer import IndexerClient
+from algopy_testing import AlgopyTestContext, algopy_testing_context
 
-from smart_contracts.artifacts.hello_world.client import HelloWorldClient
+from smart_contracts.hello_world.contract import HelloWorld
 
 
-@pytest.fixture(scope="session")
-def hello_world_client(
-    algod_client: AlgodClient, indexer_client: IndexerClient
-) -> HelloWorldClient:
-    config.configure(
-        debug=True,
-        # trace_all=True,
-    )
-
-    client = HelloWorldClient(
-        algod_client,
-        creator=get_localnet_default_account(algod_client),
-        indexer_client=indexer_client,
-    )
-
-    client.deploy(
-        on_schema_break=algokit_utils.OnSchemaBreak.AppendApp,
-        on_update=algokit_utils.OnUpdate.AppendApp,
-    )
-    return client
+@pytest.fixture()
+def context() -> Generator[AlgopyTestContext, None, None]:
+    with algopy_testing_context() as ctx:
+        yield ctx
+        ctx.reset()
 
 
-def test_says_hello(hello_world_client: HelloWorldClient) -> None:
-    result = hello_world_client.hello(name="World")
+def test_hello(context: AlgopyTestContext) -> None:
+    # Arrange
+    dummy_input = context.any_string()
+    contract = HelloWorld()
 
-    assert result.return_value == "Hello, World"
+    # Act
+    output = contract.hello(dummy_input)
 
-
-def test_simulate_says_hello_with_correct_budget_consumed(
-    hello_world_client: HelloWorldClient, algod_client: AlgodClient
-) -> None:
-    result = (
-        hello_world_client.compose().hello(name="World").hello(name="Jane").simulate()
-    )
-
-    assert result.abi_results[0].return_value == "Hello, World"
-    assert result.abi_results[1].return_value == "Hello, Jane"
-    assert result.simulate_response["txn-groups"][0]["app-budget-consumed"] < 100
+    # Assert
+    assert output == f"Hello, {dummy_input}"
