+++ template_content/smart_contracts/helpers/util.py
