+++ template_content/{% if use_python_pytest %}tests{% endif %}/{% if use_python_pytest %}{{ contract_name }}_test.py{% endif %}.jinja
@@ -1,48 +1,25 @@
-import algokit_utils
+from collections.abc import Generator
+
 import pytest
-from algokit_utils import get_localnet_default_account
-from algokit_utils.config import config
-from algosdk.v2client.algod import AlgodClient
-from algosdk.v2client.indexer import IndexerClient
+from algopy_testing import AlgopyTestContext, algopy_testing_context
 
-from smart_contracts.artifacts.{{ contract_name }}.client import {% include pathjoin('includes', 'contract_name_pascal.jinja') %}Client
+from smart_contracts.{{ contract_name }}.contract import {{ contract_name.split('_')|map('capitalize')|join }}
 
 
-@pytest.fixture(scope="session")
-def {{ contract_name }}_client(
-    algod_client: AlgodClient, indexer_client: IndexerClient
-) -> {% include pathjoin('includes', 'contract_name_pascal.jinja') %}Client:
-    config.configure(
-        debug=True,
-        # trace_all=True,
-    )
-
-    client = {% include pathjoin('includes', 'contract_name_pascal.jinja') %}Client(
-        algod_client,
-        creator=get_localnet_default_account(algod_client),
-        indexer_client=indexer_client,
-    )
-
-    client.deploy(
-        on_schema_break=algokit_utils.OnSchemaBreak.AppendApp,
-        on_update=algokit_utils.OnUpdate.AppendApp,
-    )
-    return client
+@pytest.fixture()
+def context() -> Generator[AlgopyTestContext, None, None]:
+    with algopy_testing_context() as ctx:
+        yield ctx
+        ctx.reset()
 
 
-def test_says_hello({{ contract_name }}_client: {% include pathjoin('includes', 'contract_name_pascal.jinja') %}Client) -> None:
-    result = {{ contract_name }}_client.hello(name="World")
+def test_hello(context: AlgopyTestContext) -> None:
+    # Arrange
+    dummy_input = context.any_string()
+    contract = {{ contract_name.split('_')|map('capitalize')|join }}()
 
-    assert result.return_value == "Hello, World"
+    # Act
+    output = contract.hello(dummy_input)
 
-
-def test_simulate_says_hello_with_correct_budget_consumed(
-    {{ contract_name }}_client: {% include pathjoin('includes', 'contract_name_pascal.jinja') %}Client, algod_client: AlgodClient
-) -> None:
-    result = (
-        {{ contract_name }}_client.compose().hello(name="World").hello(name="Jane").simulate()
-    )
-
-    assert result.abi_results[0].return_value == "Hello, World"
-    assert result.abi_results[1].return_value == "Hello, Jane"
-    assert result.simulate_response["txn-groups"][0]["app-budget-consumed"] < 100
+    # Assert
+    assert output == f"Hello, {dummy_input}"
